drop the duplicate code and we'll import getUrlByParameter from scripts/utils.js
