+++ test/twitter.spec.js
@@ -1,6 +1,6 @@
 const dom = require("./setup").jsdom;
 const expect = require("chai").expect;
-const getUrlByParameter = require("../scripts/twitter").getUrlByParameter;
+const getUrlByParameter = require("../scripts/utils").getUrlByParameter;
 
 describe('twitter', () =>{
   it('should extract correct tweet url', () =>{
