+++ twitter.html
@@ -48,5 +48,6 @@
 
     </body>
 </html>
+<script src="scripts/utils.js"></script>
 <script src="scripts/twitter.js"></script>
 <script src="scripts/twitter_loader.js"></script>
